audio_svc/pyproject.toml
[project]
name = "audio-svc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "audioread==3.0.1",
    "grpcio==1.66.1",
    "grpcio-tools==1.66.1",
    "librosa==0.10.2.post1",
    "numpy==2.0.2",
    "protobuf==5.29.0",
    "scipy==1.14.1",
    "soundfile==0.12.1",
]
